# Documentacion/YAML/CineReservaOpenAPI-1.0.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: CineReservas API
  description: API para gestionar reservas de un cine.
paths: 
...
  /showtimes/{movieID}:
    get:
      summary: Obten las funciones disponibles de una película.
      operationId: getShowtimes
      parameters:
        - name: movieID
          in: path
          required: true
          description: ID de la película.
          schema:
            type: integer
      responses:
        '200':
          description: Lista de funciones.
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Showtime'
        '404':
          description: Película no encontrada.
...
components:
  schemas:
...
    Showtime:
      type: object
      properties: 
        id:
          type: integer
          example: 1
        movieID: 
          type: integer
          example: 1
        room:
          type: string
          example: Sala 1
        startTime:
          type: string
          format: date-time
          example: 2025-03-14T13:00:00
...